requirements.txt
autopep8==1.4.4
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycodestyle==2.5.0
SQLAlchemy==1.3.7
Werkzeug==0.15.5
textblob
beautifulsoup4>=4.9.3
certifi>=2019.9.11
chardet>=3.0.4
Click>=7.0
colorama>=0.4.4
cssselect>=1.1.0
feedfinder2>=0.0.4
feedparser>=6.0.8
filelock>=3.0.12
Flask>=1.1.1
gunicorn>=19.9.0
idna>=2.8
imageio>=2.4.1
itsdangerous>=1.1.0
jieba3k>=0.35.1
Jinja2>=3.0.1
joblib>=1.0.1
lxml>=4.6.3
MarkupSafe>=2.0.1
newspaper3k>=0.2.8
nltk>=3.6.2
numpy>=1.21.1
Pillow>=8.3.1
pybase64>=1.0.0
python-dateutil>=2.8.2
python-dotenv>=0.10.3
PyYAML>=5.4.1
regex>=2021.7.6
requests>=2.22.0
requests-file>=1.5.1
sgmllib3k>=1.0.0
singledispatch>=3.6.2
six>=1.16.0
soupsieve>=1.9.4
tinysegmenter>=0.3
tldextract>=3.1.0
tqdm>=4.61.2
urllib3>=1.26.6
Werkzeug>=0.16.0
wordcloud>=1.5.0
WTForms>=2.2.1
